+++ src/array/codec.rs
@@ -83,7 +83,7 @@ use std::{
 
 use super::{
     concurrency::RecommendedConcurrency, ArrayView, BytesRepresentation, ChunkRepresentation,
-    DataType, MaybeBytes,
+    ChunkShape, DataType, MaybeBytes,
 };
 
 /// A codec plugin.
@@ -252,6 +252,18 @@ pub trait ArrayCodecTraits: CodecTraits {
             decoded_offset += length;
         }
         Ok(())
+    }
+
+    /// Return the partial decode granularity.
+    ///
+    /// This represents the shape of the smallest subset of a chunk that can be efficiently decoded if the chunk were subdivided into a regular grid.
+    /// For most codecs, this is just the shape of the chunk.
+    /// It is the shape of the "inner chunks" for the sharding codec.
+    fn partial_decode_granularity(
+        &self,
+        decoded_representation: &ChunkRepresentation,
+    ) -> ChunkShape {
+        decoded_representation.shape().into()
     }
 }