+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
- - Add `byte_range::extract_byte_ranges_concat{_unchecked}()`
+ - Add `byte_range::extract_byte_ranges_concat()`
+ - Add `{Async}BytesPartialDecoderTraits::partial_decode_concat()` with default implementation
 
 ## [0.12.0] - 2024-02-22